Fixed bugs and added comments

+++ server/public/scripts/controllers/projects.controller.js
@@ -1,10 +1,15 @@
 myApp.controller('ProjectsController', function ($http, $mdDialog, $mdToast) {
     const vm = this;
+    // array for storing projects
     vm.projectList = [];
+    //boolean for changing form to edit/add
     vm.projectEditBool = false;
+    // variable for setting project input field when editing
     let projectUpdateId = 0;
-
-    //TODO
+    // makes chart hidden on projects view
+    document.getElementById("myChart").setAttribute('style', "display: none;");
+    
+    // function for adding/updating projects
     vm.addProject = function () {
         vm.projectToAdd = {
             name: vm.name
@@ -43,7 +48,7 @@ myApp.controller('ProjectsController', function ($http, $mdDialog, $mdToast) {
             })
         }
     }
-
+    //function for setting edit values in the form
     vm.editProject = function (project) {
         vm.projectEditBool = !vm.projectEditBool;
         if (vm.projectEditBool) {
@@ -54,7 +59,7 @@ myApp.controller('ProjectsController', function ($http, $mdDialog, $mdToast) {
             vm.name = '';
         }
     }
-
+    //function for GETting project information
     function getProjects() {
         $http({
             method: 'GET',
@@ -65,19 +70,38 @@ myApp.controller('ProjectsController', function ($http, $mdDialog, $mdToast) {
             console.log(error);
         });//End GET
     }
-
-    document.getElementById("myChart").setAttribute('style', "display: none;");
+    //function for removing projects
     vm.deleteProject = function (project) {
         console.log('in delete');
-        $http({
-            method: 'DELETE',
-            url: '/task/project/' + project.projectid
-        }).then(function (response) {
-            console.log('Project Deleted');
-            getProjects();
-        }).catch(function (error) {
-            console.log(error);
-        });//End DELETE
+        if(project.hours > 0 || project.minutes > 0){
+            $mdDialog.show(
+                $mdDialog.alert()
+                    .title('Cannot delete a project with entries attached!')
+                    .ok('Okay')
+            );
+        }
+        else{
+            $http({
+                method: 'DELETE',
+                url: '/task/project/' + project.projectid
+            }).then(function (response) {
+                $mdToast.show(
+                    $mdToast.simple()
+                        .textContent('Project Deleted')
+                        .hideDelay(3000)
+                );
+                console.log('Project Deleted');
+                getProjects();
+            }).catch(function (error) {
+                $mdDialog.show(
+                    $mdDialog.alert()
+                        .title('Error deleting Project!')
+                        .ok('Okay')
+                );
+                console.log(error);
+            });//End DELETE
+        }
     }
+    //call to getProjects
     getProjects();
 });
